+++ chat-client/src/helpers/jwtHelper.ts
@@ -1,0 +1,16 @@
+import { jwtDecode } from "jwt-decode";
+
+const parseJWT = (): number => {
+  const token = localStorage.getItem("token");
+
+  if (token) {
+    const decoded: { id: number } = jwtDecode(token);
+    const userID = decoded.id;
+
+    return userID;
+  }
+
+  return -1;
+};
+
+export { parseJWT };
