+++ chat-client/src/hooks/useChat.ts
@@ -5,7 +5,7 @@ import {
   sendMessage as sendMessageApi,
   getMessagesByUser,
 } from "../services/api";
-import { jwtDecode } from "jwt-decode";
+import { parseJWT } from "../helpers/jwtHelper";
 
 const socket = io("http://localhost:3001");
 
@@ -13,15 +13,11 @@ export const useChat = () => {
   const [messages, setMessages] = useState<Message[]>([]);
 
   useEffect(() => {
-    const token = localStorage.getItem("token");
-    if (token) {
-      const decoded: { id: number } = jwtDecode(token);
-      const userID = decoded.id;
+    const userID = parseJWT();
 
-      getMessagesByUser(userID).then((response) => {
-        setMessages(response.data);
-      });
-    }
+    getMessagesByUser(userID).then((response) => {
+      setMessages(response.data);
+    });
 
     socket.on("message", (message: Message) => {
       setMessages((prev) => [...prev, message]);
@@ -32,31 +28,26 @@ export const useChat = () => {
   }, []);
 
   const sendMessage = async (content: string) => {
-    const token = localStorage.getItem("token");
+    const userID = parseJWT();
 
-    if (token) {
-      const decoded: { id: number } = jwtDecode(token);
-      const userID = decoded.id;
+    const userMessage: Message = {
+      sender: "user",
+      content,
+      timestamp: new Date(),
+    };
+    setMessages((prev) => [...prev, userMessage]);
+    const response = await sendMessageApi({
+      sender: "user",
+      content,
+      userID,
+    });
 
-      const userMessage: Message = {
-        sender: "user",
-        content,
-        timestamp: new Date(),
-      };
-      setMessages((prev) => [...prev, userMessage]);
-      const response = await sendMessageApi({
-        sender: "user",
-        content,
-        userID,
-      });
-
-      const botMessage: Message = {
-        sender: "bot",
-        content: response.data,
-        timestamp: new Date(),
-      };
-      setMessages((prev) => [...prev, botMessage]);
-    }
+    const botMessage: Message = {
+      sender: "bot",
+      content: response.data,
+      timestamp: new Date(),
+    };
+    setMessages((prev) => [...prev, botMessage]);
   };
 
   return { messages, sendMessage };
